Add Edit button functionality and style update
In this commit, we added the functionality to edit a to-do item by clicking the Edit button. We replaced the previous click-on-text-to-edit functionality with a new Edit button that allows users to edit to-do items more intuitively. Additionally, we updated the Edit button's style, including a darker shade of blue and hover effect, to match the appearance and interactivity of the other buttons (Selesai and Hapus).